requirements.txt
azure-functions
aiohttp
pandas
nest_asyncio
requests